pyproject.toml
[project]
name = "agentic-ai-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
"python-dotenv>=1.0.0",
"requests>=2.28.2",
"jupyter>=1.0.0",
"notebook>=6.5.3"
]
